cleanup: deprecate some IAM types
These types are only used in (already) deprecated IAM functions. I
wanted to deprecate each type, but that requires too many changes to
disable the diagnostics where we use the types. Instead I updated the
documentation and used deprecation attributes in the functions a
customer would use if they used the type.