feat(ci): add caching to GCB builds (#6084)
Adds a ci/cloudbuild/cache.sh script that is responsible for save'ing and restore'ing our build caches to/from GCS.
There are also two new build steps in the cloudbuild.yaml file that call this script at the appropriate times.

NOTE: The ~/.cache/bazel directory is not currently cached. It was huge and it wasn't clear if it was beneficial. I'll look into
fixing that later, or just using the normal remote build cache for bazel.

This PR also creates a home directory mount of /h that's visible to all build steps so they can all read/write to the homedir.