feature(spanner): request priority
Support the marking of requests with a relative priority of high,
medium, or low.  This acts as a hint to the Cloud Spanner scheduler,
and does not actually guarantee priority or order of execution.
If unspecified, the priority defaults to high.

Introduce the `google::cloud::spanner::RequestPriority` enum, and
add one as an optional field to `CommitOptions`, `QueryOptions`,
and `ReadOptions`, with builder-pattern support where applicable.

`CommitOptions` can already be applied to `Client::Commit()`, and
`ReadOptions` to `Client::Read()` and `PartitionRead()`.

`QueryOptions` can already be applied to `Client::ExecuteQuery()`,
`ProfileQuery()`, `ExecuteDml()`, `ProfileDml()`, and `AnalyzeSql()`,
but here it is also added to `Client::ExecuteBatchDml()` and
`ExecutePartitionedDml()`.