chore: building blocks for patching googleapis after download (#6104)
When developing a new feature based on simple changes to the
googleapis repo, it may well be easier to apply a patch to the
downloaded proto files rather than dealing with branches.

Those googleapis branches are often not kept up to date, so it
can be difficult to rebase branches in this repo, which hinders
developing at head. Patching also allows for the possibility of
layering changes from multiple googleapis branches.

This PR doesn't change any behavior, but it does leave indicators
for where patches can be applied in both the CMake and bazel
builds, as well as adding the "patch" binary to the docker images.

I recently used this mechanism to check whether a proposed proto
rearrangement would affect the build for the Cloud Spanner client.
An additional lesson from that effort was that any changes to the
shape of the proto tree made by the patch, which would require edits
to the list of proto files in external/googleapis/CMakeLists.txt,
would also necessitate patching the downloaded bazel BUILD files.