refactor(ci): simplified gcb build.sh to accept --project (#6108)
This PR simplifies the `ci/cloudbuild/build.sh` script so that it now
takes an (optional) `--project` flag (the same as `gcloud`), and it no
longer requires/supports the `--cache_bucket` flag. The cache bucket is
now computed from the specified (or defaulted) project ID, and it's the
same cache bucket that Google Cloud Build uses for its source upload.
The bucket is: `${PROJECT_ID}_cloudbuild`. The nice thing is that GCB
creates this bucket if necessary automatically, and we re-use this same
bucket for our build cache and bazel cache. So we only have a single
"cloudbuild" bucket now, instead of two.

I also changed the "base" branch cache key to be named "main", since
this will eventually match our `main` branch, and that's the build that
it corresponds to. I did this now since changing the cache bucket will
have to rebuild all the caches anyway, so better to do that only once.