ci(gcb): adds a "checkers" build (#6133)
This build only runs ci/check-style.sh and ci/check-markdown.sh, but more can be added as we need them. This build does not compile the code, it's just for meta-style checks about the code, formatting, etc. It runs fast, and can be run on the code in your repo with: ci/cloudbuild/build.sh checkers --docker.

Note: I changed the check-style.sh script's printf calls to write to stderr, because time also writes to stderr, and their output was coming out out-of-order in GCB without this. But since the printf and time outputs are intended to appear on the same line, it makes sense to write them both to the same FD.

This also adds an env var, GOOGLE_CLOUD_BUILD=yes, to all builds that are running on GCB. This will let us distinguish between builds running in the cloud vs builds running --local or --docker.