ci: prevent flakes in Bazel builds
Download the dependencies of a Bazel build before starting the build,
with retries. This adds little or no cost to a normal build, as the
dependencies would have been downloaded anyway, but prevents builds from
flaking out when there is some kind of network error that prevents the
download from working.