ci(gdb): enable GCS tests that require json and p12 keyfiles
The keyfiles will be stored in a special GCS bucket that the cloudbuild
account has access to read. The keyfiles are named `key-YYYY-MM.ext` to
make it easier to rotate the keys. The idea is that GCB jobs should
always look for the key w/ the current month. Rotating the keys will be
done separately.