doc(storage): always use signing account in examples (#6149)
The storage client library defaults to the account used in the
credentials as the signing account for signed URLs. This does not work
for users running the examples in the command-line, as only service
accounts can sign urls. It also does not work for the cloud builds,
where the account used for authentication cannot sign blobs as itself,
but can sign blobs as "other" accounts.

This PR changes the examples to require a signing account parameter.
Fortunately, we already had a convenient environment variable with such
an account.