cleanup: do not duplicate the Status code within its message
When adding detail to a `Status` message, but keeping the same
code, use `status.message()` to obtain the previous text, rather
than streaming `status`, as the latter will contain " [CODE]",
which will then be duplicated when the new value is streamed.

For example, when prefixing "func: " to an existing `Status` of
"file [NOT_FOUND]", we want to produce "func: file [NOT_FOUND]"
and not "func: file [NOT_FOUND] [NOT_FOUND]".