feat(storage): reduce overhead in pooled handle factory (#6169)
Replace `std::vector<>` with `std::deque<>` because when the pool grows
too large we want to remove the *oldest* handles (fron the front). Also,
use a while loop during cleanup because it makes the intent clearer. I
expect that while loop to run at most once, but why force humans to
think this through?