refactor(storage): use g::c::Options for GrpcClient
Change `storage::internal::GrpcClient` and
`storage::internal::HybridClient` to use `google::cloud::Options`. I had
to define exactly where the defaults are applied to an `Options` object,
for now this is in the `*Client::Create()` factory functions, it may get
refined later.

Some additional refactoring in `storage::ClientOptions` and
`storage::internal::CurlClient` to share the initialization of default
options.