ci(gcb): pass PROJECT_ID and BUILD_ID to build scripts (#6222)
I'll need the PROJECT_ID in an upcoming PR. I don't need BUILD_ID yet,
but it seemed useful to pass along too.

PROJECT_ID shadowed a variable in the build.sh script, so I renamed the
flag arguments in the build.sh script to have a `_FLAG` suffix. The
`build.sh` changes are purely renames.