ci(gcb): run integration tests from sanitizer builds (#6227)
Fixes #6179
Part of #6181

The basic idea of this PR is that the builds/integration.sh script was refactored into a builds/lib/integration.sh library, and the args and the bazel_with_emulators functionality was split into two functions. This will allow us to run the integration tests with emulators and against prod (in a future PR).