ci(gcb): add demo-centos-7 install build (#6239)
* ci(gcb): add demo-centos-7 install build

Part of #6163

This is the first example of how I plan to move what we currently call
"install" builds (defined in `ci/kokoro/install`) over to GCB where I
plan to call them "demo" builds. The idea is that we'll have a single
script called `demo-install.sh` (in this PR) that installs
`google-cloud-cpp`, and then tests the install by running the
quickstarts. We'll then have a number of `demo-<os>.Dockerfile` files
(`demo-centos-7.Dockerfile` is included in this PR), and a number of
trigger files, which associate a `demo-<os>` distro with the one
`demo-install.sh` script.

In subsequent PRs, I'll add more `demo-<os>.Dockerfile`s, and more
triggers for the other platforms.

Separately, I'll also update the packaging scripts to use these new
files for generating the `packaging.md` file.

Note about naming: I thought that "install" was a slightly ambiguous
term for these builds, and thought "demo" was a little more descriptive
since we use these builds to demonstrate how to install the deps and
install our library on various platforms. Happy to change the naming
back to "install" if others dislike the new naming.

* removed unneeded export