ci(gcb): build.sh now has a --trigger flag
With the addition of #6239 and other upcoming "demo" builds, we'll have
a several trigger files defining builds that all use the same underlying
build script. For example, the demo builds will all use the same
`demo-install.sh` script. So in order to run these builds from the
command-line conveniently with `build.sh`, we need a way to say "run the
build defined by trigger X".

This PR adds the `--trigger=name` flag, which takes the name of a build
trigger file where it will extract the build script and distro.

BREAKING CHANGE (not user facing): If anyone else was already using this
script for builds, the following syntax is changing.

```
build.sh asan  # Will no longer work; Instead use
build.sh -t asan-pr
```

This PR also expands the help docs for this script to hopefully make it
easier to learn.