ci!: generator/ is no longer built by default in CMake
Our `generator/` code is an implementation that our users will rarely,
if ever, need to build or use directly. So it seems it would be better
to not compile it by default in our CMake builds. This PR changes the
generator to not be build by default in CMake builds.

It also explicitly enables the generator in our clang-tidy builds
(kokoro and GCB) so that the generator is checked. All the sanitizer
builds use bazel so they will be unaffected.

This is technically a breaking change, but it's very unlikely to affect
anyone in practice.