ci(gcb): add shared library build
Also fixed an issue where the pubsub protos were missing a needed dep.

This PR refactors the logic to run the CMake and Makefile quickstart
builds into a library, because it turns out that this is a useful way to
verify that our installed artifacts work, and we end up calling this
from a few places.