cleanup(storage): deprecate overly permissive functions (#6258)
`storage::Client` has a number of functions that expose more details of
the implementation than they should, or are only intended for tests,
benchmarks, or to implement additional functionality in the library.
This change marks these functions as deprecated, and recommends what to
do instead. We also changed our tests, examples, and implementation to
use a helper in `google::cloud::internal` to access this functionality.