fix: gcb coverage build
The coverage build [failed][failure-log] with

(error omitted because conventionalcommits bot can't handle it)

I'm not sure why that failed, but in looking into it, I noticed I was
missing some needed flags for coverage, so maybe this is the problem and
will fix it. Note, that I was unable to reproduce the failure in any of
my tests, so I'm not sure if this will fix it, but I think we need these
flags anyway.

[failure-log]: https://pantheon.corp.google.com/cloud-build/builds;region=global/5f4c6f26-4330-44bd-8415-b75c43be2276;step=2?project=cloud-cpp-testing-resources