feat(common): support unified credentials in gRPC
To use the unified authentication client credentials with gRPC we need
to map said credentials to the gRPC authentication abstractions. These
come in two flavors:

- For short-lived tokens (such as access tokens), one needs to create a
  `grpc::CallCredentials` object and pass it to `grpc::ClientContext`
  before each call.
- For long-lived credentials (e.g. `grpc::GoogleDefaultCredentials`) the
  credentials are associated with the gRPC channel and gRPC takes care
  of refreshing them.

This suggests that we need an abstraction to (a) create properly
configured channels when the unified credentials can be represented by a
`grpc::ChannelCredentials`, and (b) create (and if necessary cache) the
right `grpc::CallCredentials` and then set them before each RPC.

This change introduces the first versions of these components, and
creates a mapping function. For backwards compatibility we can also map
a `std::shared_ptr<grpc::ChannelCredentials>`.