feat(common): bootstrap Unified Authentication Client (#6299)
Create the initial classes for the "Unified Authentication Client". In
this change we create (non-public) classes to represent either (a) I
want to use the default google credentials, or (b) I want to use an
access token. Future PRs will use these preferences to initialize the
client libraries.

These two cases are intereseting because, while easy to implement, they
represent two main cases for gRPC: a `grpc::ChannelCredentials` that
remains usable for a long time, and short-lived access tokens that are
provided as `grpc::CallCredentials` in each RPC (though probably
computed only once an hour or so).