feat(storage): use the unified authentication client (#6323)
Modify the storage library to support the unified authentication client
(aka `google::cloud::{internal::}Credentials`). This is not yet exposed
in the public API, but applications will be able to initialize a
`gcs::Client` with a single credential, and it will work for both REST
and gRPC. Moreover, this supports credentials that dynamically create
access tokens, which we could not do before.

This is a backwards compatible change, using `oauth2::Credentials` (or
`grpc::ChannelCredentials` for the GCS+gRPC plugin) still works.