ci(gcb): add scan-build but no trigger (#6356)
Fixes: #6345

This is a build that we may want to run sometimes by hand, and so here's
the build script for it. At this point, there's no build trigger for
this build, though we could add one if it seems useful. For now, this
build exists so we don't forget how to run a scan-build over our code.