ci(gcb): checkers build generates markdown
Much like how the checkers build formats files and leaves the edited
files in place, the markdown files will now be automatically generated
by the checkers build. Any result diffs will result in a build break
(as before), but now the edited files will be left in place so that the
user doesn't need to do a second step of manually running the generator.

With this change, once we stop kokoro, we can also remove
`ci/check-markdown.sh`