ci: make install Dockerfiles easier to update (#6379)
I neglected these Dockefiles in my similar PR for `ci/cloudbuild`
because they are going to be retired "soon". However, we will need to
update the `doc/packaging.md` file at that time, and will need to update
the scripts to generate this document too. Seems easier to do now while
the changes are fresh in my mind.