feat: a cache for access tokens retrieved using gRPC (#6381)
To support service account impersonation we will need to retrieve access
tokens via gRPC calls (to the IAM Credentials service). Access tokens
can remain valid for minutes (up to one hour), so it is worthwhile to
cache them and avoid additional RPCs to retrieve a new token. In
addition, some of our libraries use asynchronous RPCs, so we need to
retrieve the tokens asynchronously too. Moreover, application developers
expect consistent performance, having one RPC every hour take "a long
time" because it needed to fetch the access token is less than ideal. It
would be better retrieve the access tokens in the background, before
they expire.

This change introduces a class to support these features: asynchronous
retrieval, caching, pre-fetch before it expires. It could be improved
with a timer to pre-fetch even if there are no requests, but that seems
like a more subtle tradeoff (if there are no requests, why fetch the
access token at all?), and I will leave it until later.