ci(gcb): include log links in log output (#6389)
It can be convenient to link to logs in Issues, for example if a build
fails. We have two different links to logs, either in the GCB console or
a raw log file that's publicly accessible. This change makes it easy for
us to copy-n-paste links to the a log that includes both the GCB and raw
log links.