ci: rename check-abi -> check-api (#6422)
Fixes: #2907

Note that some places I left a**b**i in the name. For example, the dump
files are really a**b**i dump files, even though we're only using them
to check for a**p**i compliance. But this rename better reflects what
we're actually trying to test with this build: that our source-level
a**p**i isn't broken.

If this mixed a**b**i vs a**p**i naming seems too confusing, I'd be
happy to drop this PR and keep the old a**b**i naming. Possibly even go
all the way the other direction toward a**p**i would be OK too, if
someone else preferred that naming.