test(common): reduce timing flakiness in async_retry_loop test (#6436)
The expectation in `AsyncRetryLoopTest.ExhaustedDuringBackoff` assumes
that the retry loop will call the functor before the retry policy is
exhausted.  This can't be (easily) guaranteed, so let's just extend
the retry time limit a little in order to increase the probability
that the functor is called at least once.

Fixes #6266.