feat(common): credentials for service account impersonation (#6429)
This introduces a unified representation for service account
impersonation credentials. Service account impersonation credentials
allow a program to run with a given set of credentials (the base
credentials), but make RPCs using the credentials of a target service
account. This is often used to delegate billing to another service
account, or to support some forms of multi-tenancy.