test(spanner): PickRandomInstance
The Spanner-emulator `InstanceAdminClient::ListInstances()` does
not heed the `filter` parameter to restrict the set of instances
included in the response.  This meant that `PickRandomInstance()`
could return an ephemeral instance created, and soon to be deleted,
by a concurrent test.  And that leads to flakiness at best.

So, repeat the name filtering of instances in responses received
from the emulator.  And now that we are adding that check, make
an unexpected name from a real server return an error.

Fixed #6471.