ci: move `cmake-oldest-deps` build to Ubuntu:20.04 (#6497)
Picking Fedora to run the `cmake-oldest-deps` build was not a great idea
(it was mine). Fedora moves fast, picking new compilers very 6 months,
and dropping support every 12 months. We want this build to use
something more stable. This change moves the build to Ubuntu:20.04
(Focal Fossa), it was released in 2020-04, should remain supported until
2025-04. Most likely we will be forced to update the minimum version for
our deps before then.

In case you are wondering: yes, the older versions of our deps do not
compile with Fedora:34.