ci: make Fedora version configurable
Change the dockerfiles to make *only* the Fedora version configurable.
The other builds have the version fixed in their name, e.g.,
`ubuntu-bionic.Dockerfile` obviously is intended to be used exclusively
with `ubuntu:bionic` (aka `ubuntu:18.04`).

Change the build.sh script to accept a `--distro-version` flag to set
the `DISTRO_VERSION` build argument in `docker build`. Also changed the
triggers to set `_DISTRO_VERSION` substitution, so this value is picked
up correctly in builds with `-t` flag. The substitution is not used in
the GCB builds yet.