ci: create dockerfiles for Fedora:33 and Fedora:34 (#6503)
Remove the `DISTRO_VERSION` from all the Dockerfiles.
For Fedora, copy the existing file to `fedora-33.Dockerfile` and `fedora-34.Dockerfile`.
Once this is merged and the builds are moved to use `fedora-33` I will send a second
PR to remove `fedora.Dockerfile`.  And subsequent PR will move some builds to
`fedora-34`.