ci: simplify and document generator formatting trick (#6514)
The trick we currently use to clang-format most (but not all) of the
generator files has a much better solution in clang-format version 13. I
created an issue so we don't forget about this and I added a TODO in the
code linking this issue. I also simplified the code a bit to just a
single cp.