ci: exclude hello_world_http from doxygen
To include this in doxygen output we'll need to fetch the functions
framework dep, which we could do, but that would be a big-ish change for
doxygen. Maybe we can consider that later, but for now, I think
excluding this from doxygen sounds better.

An alternative would be to move this `hello_world_http` somewhere else,
which may make sense.