ci: make doxygen errors break the build
Fixes: #6543

Our user-facing documentation is important and if we create docs with
broken links and what not, we should fix it. If this becomes too
cumbersome or if it has too many false positives, we can disable this
again. But I think it's worth a try to see how it goes.