test(common): eliminate flakiness in async_retry_loop test (#6585)
The expectation in `AsyncRetryLoopTest.ExhaustedDuringBackoff` assumes
that the retry loop will call the functor before the retry policy is
exhausted. Previously (in #6436) we extended the retry time limit a little
in order to increase the probability that the functor is called at least
once. Now we replace the time-based retry policy with an error-count
one, which guarantees the (failing) functor is called once.

Really fixes #6266.