ci: allocate a docker tty only when we have one
Conditionalize the docker `--tty` flag so that we only allocate
a tty inside the image when we have a tty to connect it to. The
major upshot of this is that you can now pipe shell commands into
`ci/cloudbuild/build.sh --docker --trigger=<trigger> --docker-shell`.

Previously, under the unconditional `--tty=true`, docker would set
`TERM=xterm`, which was bad when there was no external tty, and
possibly wrong when there was one.  So, we now also conditionally
set `${TERM}` within the image depending on whether there is an
external tty and what its terminal type is, and otherwise default
to `TERM-dumb`. This should eliminate the xterm control sequences
we see in log files (and correct things for those who do not use
an xterm-compatible terminal).