feat: custom trust store with unified credentials
Some applications may not be able to use the system's trust store (aka
CA certificates, aka root of trust files). For example, because the
application is running in a lightweight container without system
configuration files. And because the credentials need to use this trust
store to create and/or refresh access tokens, we need a unified
configuration parameter to set the trust store.

This PR introduces `google::cloud::CAInfo` an option that should point
to a file with the CA certificate bundle (typically a PEM file). For
REST-based libraries we simply point libcurl (and indirectly OpenSSL) to
this file. For gRPC-based library we load the contents of this file and
feed them to gRPC.