ci: use typos-cli instead of cspell
Typos (https://crates.io/crates/typos) is a modern source-code spell
checker that is crazy fast and seems to work very well out of the box.
By default it checks all of our code, respecting our `.gitignore`,
including source code itself (not just comments). On my laptop `cspell`
took ~30 seconds and `typos` takes ~0.2 seconds, which is convenient
because most of us run this build frequently to fix formatting, etc.

It also supports the ability to write the changes back to disk
(`--write-changes`), which we could consider as we gain confidence in
it. I used this mode to fix all the typos in this PR, but it had one
case of bad changes where it changed `Expection` -> `Exception`, whereas
in the particular context we wanted `Expectation`. I tweaked this by
hand, and decided that simply reporting the misspellings rather than
fixing them is probably safer, and it matches what we did w/ cspell.