ci: split checkers build to its own Dockerfile
This is useful because the "checkers" build is unique in that it doesn't
actually compile any code, rather it simply checks things _about_ the
code, such as formatting, whitespace, etc. Therefore, the docker
container in which it runs may likely need to be quite different than
normal builds which will need compilers, etc. To keep a bit separate and
to minimize unneeded cross-dependencies, it seems reasonable to split
this build out into its own dockerfile.