fix(cmake-oldest-deps): use vcpkg release tarball rather than clone
Fixes: https://github.com/googleapis/google-cloud-cpp/issues/6699

This PR changes the `cmake-oldest-deps` build to use a release tarball
of vcpkg, the same as our other builds that use vcpkg, instead of
cloning the whole vcpkg repo, which picks up new changes from vcpkg that
can break our build unexpectedly.

This change appears to work in my local tests, but I'm concerned that
it's wrong because the previous code had a comment saying that cloning
the repo was required. So I suppose this PR is just for discussion to
see if this is the right direction or if we should fix the issue some
other way.