ci(vcpkg): create lib/vcpkg.sh used by several builds (#6701)
Fixes: https://github.com/googleapis/google-cloud-cpp/issues/6699

This PR factors the code to fetch the vcpkg repo into a library that's
reused in a couple places. It also clones the repo at a recent release
tag rather than cloning at HEAD which may break our builds and is not
hermetic, or downloading a tarball, which lacks git history.