test(spanner): relax emulator SPANNER_STATISTICS expectation
I did not track down exactly why, but the error code from the
emulator when reading the INFORMATION_SCHEMA.SPANNER_STATISTICS
table has changed from `kInvalidArgument` in the Cloud SDK
build to `kUnimplemented` in an emulator built from source.
So, allow both.

We could be even less restrictive (like `Not(IsOk())`), but
being as exacting as possible allows us to learn of changes
in emulator behavior. For example, in some places we still
`if (emulator) GTEST_SKIP()`, but that will never let us know
when the emulator starts supporting a feature.