ci: disable test result cache on CI builds
@coryan observed that our flakiness data from looking at CI builds only
was not accurate for bazel builds because the PR builds would have
already populated the cache, and the following CI build would just hit
the cache and not even run the test, thus hiding potential flakes.

This PR avoids that issue by forcing CI builds to ignore the
test-results cache with `--cache_test_results=no`.