test(storage): resumable uploads abort on errors (#6790)
This new test verifies that resumable uploads abort (do not finalize) if
one of the intermediate `UploadChunk()` requests has a permanent error
or otherwise exhaust the retry policy. I wrote this test expecting that
the library had a bug and needed fixing, but turns out I do not give
myself enough credit.