test(storage): deflake integration tests
Use a different instance of the emulator to run the **emulator** tests
vs. the tests that **use** the emulator. The emulator tests
intentionally crash some of the emulator workers, which then causes
problems for the regular tests. Note that the emulator tests themselves
are still flaky, but they are very fast so using `--flake-test_attempts`
mostly workarounds the problem.