ci: increase the GCB timeout
We've been seeing some timeouts from the `integration-daily`
build since disabling CI-build test-result caching.  This
should not be too surprising given that the slow Spanner
backup/instance tests and samples already take a good part
of the allotted 1h, and are subject to scheduling delays in
the backend.  So, increase the timeout to 2h.

I couldn't, for the life of me, figure out how to parameterize
the timeout with a `gcloud builds submit --substitutions=`,
and chose not to move the value out to a `--timeout=` flag,
so this simply raises the limit for all builds.  Alternative
ideas are welcomed.