feat(common): a generic long-running operation
Long-running operations [aip/151] are used for API methods that take a
significant amount of time to complete (think minutes, maybe an hour).
The gRPC API returns a "promise" object, represented by the
`google::longrunning::Operation` proto, and the application (or client
library) should periodically poll this object until it is "done".

In the C++ client libraries we represent these long-running operations
by a member function that returns `future<StatusOr<ReturnType>>`. This
PR introduces a generic function to implement these member functions. It
first starts the operation using an asynchronous retry loop, and then
starts an asynchronous loop to poll the operation until it completes.

The promise can complete with an error, which is represented by a
`google::cloud::Status` object, or with success and some `ReturnType`
value.  The application may also configure the "polling policy", which
may stop the polling even though the operation has not completed.

A future PR will change the generated libraries to use this function.
With some work we could change the Spanner library to use it too,
without the need to launch detached threads to poll the results. With
substantial work we could change the Bigtable library to use it too.

[aip/151]: https://google.aip.dev/151