ci: increase the GCB timeout (#6795)
We've been seeing some timeouts from the `integration-daily`
build since disabling CI-build test-result caching.  This
should not be too surprising given that the slow Spanner
backup/instance tests and samples already take a good part
of the allotted 1h, and are subject to scheduling delays in
the backend.  So, increase the timeout to 2h.