feat(generator): use async for long-running operations
For long-running operations we need to generate three different RPCs:
one to start the operation (e.g. `CreateDatabase()`), one to poll the
operation (`GetOperation()`), and one to cancel the operation
(`CancelOperation()`).  With this change all the three RPCs are
generated as asynchronous RPCs.

The `*Connection` class can use these asynchronous RPCs in the
background, using the background threads. This (1) avoids creating a
detached thread to poll the operation, (2) uses code that is more easily
unit tested separate from the generated libraries.