chore(deps): upgrade pkg-config on CentOS-7 to avoid bug
There's a [bug] w/ CentOS-7's version of `pkg-config` that can cause
invocations to take a reeeeeaaaaly long time to complete. We discovered
this when upgrading to grpc-1.38.1 and had to [rollback] the upgrade.

The fix to `pkg-config` was rolled out about 8 years ago in `pkg-config`
0.28. This PR upgrades the pkg-config we use on CentOS-7 to the latest
as of this writing. After this, I think we should be able to upgrade
grpc again.

[bug]: https://bugs.freedesktop.org/show_bug.cgi?id=54716
[rollback]: https://github.com/googleapis/google-cloud-cpp/pull/6840