cleanup(ci): dedupe integration envvars (#6851)
This PR refactors a number of integration-test-related environment
variables so they are set consistently in a single place. This dedupes a
number of variables that were already being set.

This should also fix some flakes because we were accidentally running
the "slow" (backup and instance) integration tests more than we should:
Those tests are only supposed to be run on the daily build and code
coverage, but we were running them on all bazel builds.