fix(common): synchronize cancel for long-running operations (#6855)
Cancellation of a long-running operation (1) happens asynchronously
with the flow of callbacks, and (2) requires access to the currently
running operation name. This means we must serialize reads/writes of
that name, so introduce a member mutex.

And given that we now have a mutex, use it to guard `canceled_`
as well, avoiding the need for atomic operations. Also mark the
operation as cancelled when that happens before the operation name
is first set.

Finally, update the `DatabaseAdminConnectionTest.CreateBackupCancel`
test to allow for `AsyncCancelOperation()` not being called before
the polling policy expires.

Fixes #6854.