fix(storage): crashes with tracing enabled
With some MSVC configurations (2019+Debug, 2017+Debug, or 2017+Release)
enabling HTTP tracing creates a crash. I think because in other
compilers the move constructor and assignments are not used, and luckily
that means the pointers were stable, but this is not the case for those
versions/configurations of MSVC.

The integration test crashes before this change and succeeds after the
change.