cleanup(storage): run benchmarks with fewer privileges
The benchmarks were creating buckets with `PredefinedAcl::Private()`,
that is, a bucket that requires project owner privileges to create
objects in it. This changes the bucket permissions to
`PredefinedAcl::ProjectPrivate()`, that is, a bucket where project
editors can create objects.

I also changed the benchmarks to use the `PredefinedAcl::*()` helper
functions, they are there to prevent misspelling, might as well use
them.