cleanup(common): avoid spurious warnings
With gRPC, one must finish a streaming RPC before destroying it, which
also requires capturing the output. We do this in the destructor, and
print a warning if some error goes undetected. However, `kCancelled` is
not an unexpected error, the destructor cancels the RPC, there is no
reason to print anything in this case.