fix(GCS+gRPC): resumable uploads can resume (#6881)
To make resumable uploads resumable we need to close the stream after
each chunk, and the byte accounting needed some fixing. I was able to
simplify some of the code after these changes.

In the process I discovered and fixed a bug in the emulator: it did not
handle partial uploads correctly.