feat(GCS+gRPC): capture streaming RPCs metadata
gRPC returns metadata for all requests, including streaming RPCs. In the
storage service this is very useful for troubleshooting, this change
weaves the metadata (and some additional bits from
`grpc::ClientContext`) through the different layers. The metadata is
exposed as "headers", because that matches the existing APIs (and gRPC
uses HTTP/2 headers under the hood anyway).