ci(spanner): run spanner integration tests against prod
Add a hook for running integration tests without the emulator,
execute the Spanner tests from it, and call the hook during
the `integration-daily` build.

The Spanner samples/tests have cases that are skipped or
short-circuited when running against the emulator, so we lost
some coverage when we move to using the emulator exclusively
from the "normal" builds.  Now we'll at least hit those cases
during `integration-daily`.