fix(storage): treat 304 errors as kFailedPrecondition (#6897)
GCS returns a 304 (Not Modified) for downloads with a
`ifMetagenerationNotMatch` pre-condition that is not met.  Methinks that
is a confusing error to return ("not modified" for a read-only operation
that was not going to modify anything? wut?).  In any case, we need to
map this to our error namespace, and `kFailedPrecondition` seems like
the best mapping.

I added integration tests for the failure and success cases of each
pre-condition, so far we only had unit tests and obviously we were
missing some problems. I also added some samples because we got
questions about how to do this.