ci(spanner): run spanner integration tests against prod (#6888)
Execute the Spanner integration tests against prod from the
`integration-daily` build.

The Spanner samples/tests have cases that are skipped or
short-circuited when running against the emulator, so we lost
some coverage when we move to using the emulator exclusively
from the "normal" builds.  Now we'll at least hit those cases
during `integration-daily`.