fix(storage): pre-conditions are too restrictive
Some of our helper functions (`ParallelFileUpload()`,
`DeleteObjectsByPrefix()`) used pre-conditions to avoid deleting new
versions of an object. In most cases that is too restrictive, it
prevents acting on an object that has been modified, even if the version
we want is still available. We can use `storage::Generation()` to just
delete the version we want, even if there is a newer version.