feat(storage): benchmark aggregate download banwidth (#6927)
A new benchmark for aggregate download bandwidth. The existing
benchmarks perform both uploads and downloads as part of the run, which
limits the total download bandwidth. Moreover, their output is per
download, which makes it hard to aggregate the results across multiple
processes. This benchmark does not suffer from those problems, but it is
more limited in that it just measures one configuration at a time.