cleanup(storage): CurlHandle::SetOption assertions (#6944)
The `CurlHandle::SetOption()` functions only fail if there is a bug in the
library code, or some memory corruption that makes the library unusable.
Error handling in this case is of little use, this change renames the
error handling functions to express that they are: assertions. A future
PR will change them from throwing exceptions to calling
`GCP_LOG(FATAL)`, but that depends on fixing the handling of `FATAL`
too.