feat(storage): more accurate aggregate bandwidth results
In the aggregate_throughput benchmark we do not want to sleep for longer
than the total running time, or the main thread would be sleeping while
no work is done by the worker threads, and the elapsed time measurement
would include time when no work was being done.

Likewise, we want to wait until all threads are done with their
downloads before reporting the aggregate bandwidth.