refactor(spanner): move Base64 codecs to common (#6954)
I think these are better than the OpenSSL-based ones in the storage
library. In this step I refactored the core functions to the common
libraries.

I changed the encoder class to make it easier / safer to use outside the
`spanner::Bytes` class.

I moved most of the tests too, I did not think it made sense to
duplicate the tests.