fix(GCS+gRPC): no hashes in Object resource (#6963)
The proto file says that neither the `md5_hash` nor the `crc32c` fields
should be set when uploading an `Object`:

https://github.com/googleapis/googleapis/blob/1eade8161f2ef102282869bc3eb18841807b2b46/google/storage/v1/storage_resources.proto#L691-L694
https://github.com/googleapis/googleapis/blob/1eade8161f2ef102282869bc3eb18841807b2b46/google/storage/v1/storage_resources.proto#L676-L679

Also improved the documentation on how to obtain these hashes for tests.

I disabled some tests that used to pass because the service does not ignore
the hashes in the `Object` resource. The documentation says it does (see
the links), so my tests were passing by accident and the service can change
at any time.  Once I fix #4156 and #4157 these tests can be enabled again.